Move build package up one level
